pyproject.toml
[tool.poetry]
name = "seminar-2022-summary-model"
version = "0.1.0"
description = ""
authors = ["kathmandu777 <mahhakatomanato@gmail.com>"]
readme = "README.md"
packages = [{include = "seminar_2022_summary_model"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
transformers = "^4.25.1"
python-box = "^6.1.0"
pytorch-lightning = "^1.8.6"
scikit-learn = "^1.2.0"
wandb = "^0.13.8"
sumeval = "^0.2.2"
mecab-python3 = "^1.0.6"
sacrebleu = {version = "<2.0.0", extras = ["ja"]}
unidic-lite = "^1.0.8"
sentencepiece = "^0.1.97"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
